+++ lib/shared/models/auth_model.dart
@@ -1,7 +1,6 @@
 class UserData {
   final String email;
   final String uuid;
-  String? localDeviceId;
 
-  UserData({required this.email, required this.uuid, this.localDeviceId});
+  UserData({required this.email, required this.uuid});
 }