+++ lib/shared/widgets/account_card.dart
@@ -1,7 +1,13 @@
+import 'package:easy_localization/easy_localization.dart';
 import 'package:flutter/material.dart';
-import 'package:lucide_icons/lucide_icons.dart';
 
-Widget accountCard(BuildContext context, String email, String uuid) {
+Widget statusCard(
+  BuildContext context, {
+  required String title,
+  required String subtitle,
+  required IconData icon,
+  bool? active,
+}) {
   final tt = Theme.of(context).textTheme;
 
   return Card.outlined(
@@ -10,14 +16,32 @@ Widget accountCard(BuildContext context, String email, String uuid) {
       child: Row(
         spacing: 16,
         children: [
-          const Icon(LucideIcons.user),
+          Icon(icon),
           Column(
             crossAxisAlignment: CrossAxisAlignment.start,
             mainAxisSize: MainAxisSize.min,
             children: [
-              Text(email, style: tt.titleMedium?.copyWith(fontWeight: FontWeight.bold)),
+              Row(
+                crossAxisAlignment: CrossAxisAlignment.baseline,
+                textBaseline: TextBaseline.alphabetic,
+                spacing: 4,
+                children: [
+                  Text(title, style: tt.titleMedium?.copyWith(fontWeight: FontWeight.bold)),
+                  if (active != null) ...[
+                    const SizedBox(width: 4),
+                    Text(
+                      '•',
+                      style: TextStyle(
+                        color: active ? Colors.green : Colors.red,
+                        fontSize: tt.titleMedium?.fontSize ?? 20,
+                      ),
+                    ),
+                    Text(active ? 'online'.tr() : 'offline'.tr(), style: TextStyle(color: Colors.grey)),
+                  ],
+                ],
+              ),
               Text(
-                uuid,
+                subtitle,
                 style: tt.labelSmall?.copyWith(
                   fontFamily: 'Oxanium',
                   color: Colors.grey,