+++ lib/shared/widgets/device_binding_field.dart
@@ -2,14 +2,16 @@ import 'package:easy_localization/easy_localization.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:lucide_icons/lucide_icons.dart';
+import 'package:silversole/core/data/save_service.dart';
+import 'package:silversole/core/data/sole_service.dart';
 import 'package:silversole/core/error/error_logger.dart';
 import 'package:silversole/core/error/result.dart';
-import 'package:silversole/core/data/save_service.dart';
 import 'package:silversole/core/utils/field_validator.dart';
 import 'package:silversole/shared/providers/auth_provider.dart';
 import 'package:silversole/shared/providers/sole_provider.dart';
 
 import '../models/auth_model.dart';
+import '../providers/settings_provider.dart';
 
 class DeviceBindingField extends ConsumerStatefulWidget {
   const DeviceBindingField({super.key});
@@ -20,6 +22,7 @@ class DeviceBindingField extends ConsumerStatefulWidget {
 
 class _DeviceBindingFieldState extends ConsumerState<DeviceBindingField> {
   final formKey = GlobalKey<FormState>();
+  final controller = TextEditingController();
   bool autoValidate = false;
   bool isBinding = false;
   String deviceId = '';
@@ -38,16 +41,14 @@ class _DeviceBindingFieldState extends ConsumerState<DeviceBindingField> {
       final soleService = ref.read(soleProvider);
       final result = await soleService.bindingDevice(userProvider.uuid, deviceId);
       switch (result) {
-        case Ok():
+        case Ok<BindingResult>():
+          final message = result.value == BindingResult.alreadyBound
+              ? 'device_already_binding'.tr()
+              : 'binding_success'.tr();
           await saveDeviceId(deviceId);
-          ref.read(authUserProvider.notifier).setUser(
-                UserData(
-                  email: userProvider.email,
-                  uuid: userProvider.uuid,
-                  localDeviceId: deviceId,
-                ),
-              );
-          showMessage('binding_success'.tr());
+          ref.read(authUserProvider.notifier).setUser(UserData(email: userProvider.email, uuid: userProvider.uuid));
+          ref.read(settingsProvider.notifier).setDeviceId(deviceId);
+          showMessage(message);
         case Error():
           showErrorSnakeBar(result.error.toString());
       }
@@ -58,6 +59,21 @@ class _DeviceBindingFieldState extends ConsumerState<DeviceBindingField> {
         setState(() => isBinding = false);
       }
     }
+  }
+
+  @override
+  void initState() {
+    super.initState();
+    final settings = ref.read(settingsProvider);
+    deviceId = settings.deviceId ?? '';
+    controller.text = deviceId;
+    debugPrint('[DBF] initState');
+  }
+
+  @override
+  void dispose() {
+    controller.dispose();
+    super.dispose();
   }
 
   @override
@@ -72,6 +88,7 @@ class _DeviceBindingFieldState extends ConsumerState<DeviceBindingField> {
           children: [
             Expanded(
               child: TextFormField(
+                controller: controller,
                 keyboardType: TextInputType.emailAddress,
                 decoration: InputDecoration(
                   labelText: 'silversole_device_id'.tr(),